--- client/src/components/AddUser/AddUser.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-const AddUser = () => {
-  return (
-	<div>Add User</div>
-  );
-};
-
-export default AddUser;
