+++ client/src/components/EditUser/EditUser.css
@@ -1,0 +1,87 @@
+.signup-page {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 100%;
+}
+
+.signup-page h1 {
+  font-size: 2rem;
+  margin-bottom: 2rem;
+  color: #000;
+}
+
+.form-group {
+  margin-bottom: 1rem;
+  width: 100%;
+}
+
+.form-group label {
+  display: block;
+  margin-bottom: 0.5rem;
+}
+
+.form-group input,
+.form-group select {
+  width: 100%;
+  padding: 0.5rem;
+  border: 1px solid #ccc;
+  border-radius: 0.25rem;
+  font-size: 1rem;
+  line-height: 1.5;
+}
+
+.form-group input:focus,
+.form-group select:focus {
+  outline: none;
+  border-color: #0077ff;
+  box-shadow: 0 0 0 0.125em rgba(0, 119, 255, 0.25);
+}
+
+.language {
+  margin-bottom: 0.5rem;
+}
+
+.signup,
+.cancel {
+  padding: 0.5rem 1rem;
+  border: none;
+  border-radius: 0.25rem;
+  font-size: 1rem;
+  line-height: 1.5;
+  cursor: pointer;
+}
+
+.signup {
+  background-color: #0077ff;
+  color: #fff;
+  margin-right: 0.5rem;
+}
+
+.cancel {
+  background-color: #fff;
+  color: #000;
+}
+
+.login_button {
+  margin-top: 1rem;
+  font-size: 1rem;
+  line-height: 1.5;
+  color: #0077ff;
+  cursor: pointer;
+}
+
+.signup-page {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin: 0 auto;
+  max-width: 600px;
+  padding: 20px;
+
+  /* Media query for screens with max width of 768px */
+  @media only screen and (max-width: 768px) {
+    max-width: 90%;
+  }
+}
