+++ client/src/components/SideBar/SideBar.js
@@ -17,11 +17,11 @@ const items = [
 	{ key: "dashboard", icon: <DashboardOutlined />, label: "Dashboard" },
 	{ key: "profile", icon: <ProfileOutlined />, label: "Profile" },
 	{
-		key: "users",
+		key: "user",
 		icon: <UserSwitchOutlined />,
-		label: "Users management",
+		label: "User management",
 		children: [
-			{ key: "add-users", icon: <UsergroupAddOutlined />, label: "Add users" },
+			{ key: "edit-users", icon: <UsergroupAddOutlined />, label: "Edit users" },
 			{
 				key: "user-list",
 				icon: <OrderedListOutlined />,
